Rename drivers/driver_ug2864hsweg01.[ch] to oled_ug2864hsweg.[ch]
