+++ src/main/io/dashboard.c
@@ -29,7 +29,7 @@
 #include "build/build_config.h"
 
 #include "drivers/system.h"
-#include "drivers/display_ug2864hsweg01.h"
+#include "drivers/oled_ug2864hsweg01.h"
 
 #include "common/printf.h"
 #include "common/maths.h"
